+++ lib/simple_crud/base_controller.rb
@@ -1,4 +1,4 @@
-require_relative 'helper/model_helper'
+require_relative 'helper/resource_helper'
 require_relative 'helper/decorator_helper'
 
 module SimpleCrud
@@ -6,14 +6,14 @@ module SimpleCrud
     include ModelHelper
     include DecoratorHelper
 
-    before_filter :find_model, :only => [:show, :edit, :update, :destroy]
+    before_filter :find_resource, :only => [:show, :edit, :update, :destroy]
     respond_to :html, :json
 
     class << self
-      attr_accessor :model_klass
+      attr_accessor :resource_klass
 
       def crud_for(klass)
-        @model_klass = klass
+        @resource_klass = klass
       end
 
       def default_crud
@@ -23,83 +23,83 @@ module SimpleCrud
       end
     end
 
-    # GET /models
-    # GET /models.json
+    # GET /resources
+    # GET /resources.json
     def index
-      models_set model_klass.all
-      respond_with models
+      resources_set model_klass.all
+      respond_with resources_get
     end
 
-    # GET /models/1
-    # GET /models/1.json
+    # GET /resources/1
+    # GET /resources/1.json
     def show
-      respond_with model
+      respond_with resource_get
     end
 
-    # GET /models/new
-    # GET /models/new.json
+    # GET /resources/new
+    # GET /resources/new.json
     def new
-      model_set model_klass.new
-      respond_with model
+      resource_set model_klass.new
+      respond_with resource_get
     end
 
-    # GET /models/1/edit
+    # GET /resources/1/edit
     def edit
     end
 
-    # POST /models
-    # POST /models.json
+    # POST /resources
+    # POST /resources.json
     def create
-      model_set model_klass.new(model_params)
+      resource_set model_klass.new(model_params)
 
       respond_to do |wants|
-        result = model.save
+        result = resource_get.save
         call_hook :after_save, result
         if result
-          flash[:notice] = t 'messages.record_created', model: t("models.#{model_name}")
-          wants.html { redirect_to(model) }
-          wants.json  { render :json => model, :status => :created, :location => model }
+          flash[:notice] = t 'messages.record_created', resource: t("models.#{model_name}")
+          wants.html { redirect_to(resource_get) }
+          wants.json  { render :json => resource_get, :status => :created, :location => model }
         else
           wants.html { render :action => "new" }
-          wants.json  { render :json => model.errors, :status => :unprocessable_entity }
+          wants.json  { render :json => resource_get.errors, :status => :unprocessable_entity }
         end
       end
     end
 
-    # PUT /models/1
-    # PUT /models/1.json
+    # PUT /resources/1
+    # PUT /resources/1.json
     def update
       respond_to do |wants|
-        result = model.update_attributes(model_params)
+        result = resource_get.update_attributes(model_params)
         call_hook :after_update_attributes, result
         if result
-          flash[:notice] = t 'messages.record_updated', model: t("models.#{model_name}")
-          wants.html { redirect_to(model) }
+          flash[:notice] = t 'messages.record_updated', resource: t("models.#{model_name}")
+          wants.html { redirect_to(resource_get) }
           wants.json  { head :ok }
         else
           wants.html { render :action => "edit" }
-          wants.json  { render :json => model.errors, :status => :unprocessable_entity }
+          wants.json  { render :json => resource_get.errors, :status => :unprocessable_entity }
         end
       end
     end
 
-    # DELETE /models/1
-    # DELETE /models/1.json
+    # DELETE /resources/1
+    # DELETE /resources/1.json
     def destroy
-      result = model.destroy
+      result = resource_get.destroy
       call_hook :after_destroy, result
-      flash[:notice] = t 'messages.record_destroyed', model: t("models.#{model_name}")
+      flash[:notice] = t 'messages.record_destroyed', resource: t("models.#{model_name}")
 
       respond_to do |wants|
-        wants.html { redirect_to(models_path) }
+        wants.html { redirect_to(resources_path) }
         wants.json  { head :ok }
       end
     end
 
     private
 
-    def find_model
-      model_set model_klass.find(params[:id])
+    def find_resource
+      resource_set model_klass.find(params[:id])
     end
 
     def call_hook(method, *args)