+++ CHANGELOG.md
@@ -14,3 +14,9 @@ This file is used to list changes made in each version of the simplecrud gem.
 - Added `models_path` method
 - Fixed destroy action on BaseController
 - Added internationalization messages
+
+0.1.2
+-----
+- Renamed all reference to `model` to `resource`
+- Renamed all reference to `resource` to `resource_get`
+- Renamed all reference to `resources` to `resources_get`